parallel_downloader/src/requirements.txt
pillow==5.1.0
requests==2.18.4
